Use tree query parameter to limit jenkins load
Previously, vmstatus would retrieve all of the data for the Job and
all of the data for the specified Run (either a specific build id or
the last successful one).

This commit uses the `tree` query parameter to limit the amount of data
that needs to be generated and sent. It makes the queries faster[1], and
hopefully reduces the amount of memory consumed by Jenkins, since it
shouldn't need to load all of the Runs associated with the Job.

Also add an HTTP user-agent string to identify what agents are hitting
Jenkins.

[1] https://www.cloudbees.com/blog/taming-jenkins-json-api-depth-and-tree